feat(aoi): wire up SHAPE/POINTS/ASSET methods and extract AdminLevelSelector
- Dispatch SHAPE/POINTS/ASSET in process_aoi to the existing processors
  instead of raising "not available yet"
- Extract AdminLevelSelector into pysepal/solara/components/inputs/
  (re-exported from aoi package for back-compat)
- shape.process_shape: strip Z coords via shapely.force_2d so KML files
  don't break ee.FeatureCollection with "Invalid GeoJSON geometry"
- Use GeomanDrawControl.to_json() instead of the removed get_data()
- Clear shape/points/asset reactives on method change to avoid stale output
- Drop the "Use the drawing tools..." banner; the existing empty-features
  error already guides the user
- Trim docstring fluff and noisy inline comments from state declarations
