+++ pysepal/solara/components/inputs/__init__.py
@@ -1,11 +1,13 @@
 """Solara input components for pysepal."""
 
+from .admin_selector import AdminLevelSelector
 from .asset_select import AssetSelectComponent
 from .file_input import FileInputComponent
 from .point_selector import PointsSelectorComponent
 from .vector_selector import VectorSelectorComponent
 
 __all__ = [
+    "AdminLevelSelector",
     "FileInputComponent",
     "AssetSelectComponent",
     "PointsSelectorComponent",
